# Controllers/StopsController.cs
using System;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using Swashbuckle.AspNetCore.Annotations;
using TransportDataService;

namespace ulasım_veri_servisi.Controllers
{
    [ApiController]
    [Route("api/v1/stops")]
    public class StopsController : ControllerBase
    {
        private readonly AppDbContext _context;

        public StopsController(AppDbContext context)
        {
            _context = context;
        }

        [HttpGet]
        [SwaggerOperation(
    Summary = "Durakları listeler",
    Description = "Sayfalama ve arama desteğiyle durak listesini döndürür."
)]
        [ProducesResponseType(StatusCodes.Status200OK)]
        public IActionResult GetStops(
            string? search,
            int page = 1,
            int pageSize = 20)


        {
            var query = _context.Stops
                .Include(x => x.StopRoutes)
                .AsQueryable();

            if (!string.IsNullOrEmpty(search))
            {
                query = query.Where(x => x.Name.Contains(search));
            }

            var totalCount = query.Count();

            var items = query
     .Skip((page - 1) * pageSize)
     .Take(pageSize)
     .Select(x => new
     {
         id = x.Id,
         externalStopId = x.ExternalStopId,
         name = x.Name,
         latitude = x.Latitude,
         longitude = x.Longitude,
         routes = x.StopRoutes.Select(r => r.RouteNumber).ToList()
     })
     .ToList();

            return Ok(new
            {
                items,
                page,
                pageSize,
                totalCount
            });
        }
        [HttpGet("{id}")]
        [SwaggerOperation(
    Summary = "Id'ye göre durak getirir",
    Description = "Veritabanındaki Id değerine göre durak bilgisini döndürür."
)]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status404NotFound)]
        public IActionResult GetStopById(int id)
        {
            var stop = _context.Stops
                .Include(x => x.StopRoutes)
                .FirstOrDefault(x => x.Id == id);

            if (stop == null)
            {
                return NotFound();
            }

            return Ok(new
            {
                id = stop.Id,
                externalStopId = stop.ExternalStopId,
                name = stop.Name,
                latitude = stop.Latitude,
                longitude = stop.Longitude,
                routes = stop.StopRoutes.Select(x => x.RouteNumber).ToList()
            });
        }
        [HttpGet("by-external-id/{externalStopId}")]
        [SwaggerOperation(
    Summary = "Gerçek durak numarasına göre durak getirir",
    Description = "ESHOT durak numarasına göre durak bilgisini döndürür."
)]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status404NotFound)]
        public IActionResult GetStopByExternalId(string externalStopId)
        {
            var stop = _context.Stops
                .Include(x => x.StopRoutes)
                .FirstOrDefault(x => x.ExternalStopId == externalStopId);

            if (stop == null)
            {
                return NotFound();
            }

            return Ok(new
            {
                id = stop.Id,
                externalStopId = stop.ExternalStopId,
                name = stop.Name,
                latitude = stop.Latitude,
                longitude = stop.Longitude,
                routes = stop.StopRoutes.Select(x => x.RouteNumber).ToList()
            });
        }
        [HttpGet("nearby")]
        [SwaggerOperation(
    Summary = "Yakındaki durakları getirir",
    Description = "Verilen koordinata yakın durakları Haversine formülü ile hesaplar."
)]
        [ProducesResponseType(StatusCodes.Status200OK)]
        public IActionResult GetNearbyStops(
    double latitude,
    double longitude,
    double radiusMeters)
        {
            var items = _context.Stops
                .ToList()
                .Select(stop => new
                {
                    id = stop.Id,
                    externalStopId = stop.ExternalStopId,
                    name = stop.Name,
                    latitude = stop.Latitude,
                    longitude = stop.Longitude,
                    distanceMeters = CalculateDistance(
                        latitude,
                        longitude,
                        stop.Latitude,
                        stop.Longitude)
                })
                .Where(x => x.distanceMeters <= radiusMeters)
                .OrderBy(x => x.distanceMeters)
                .ToList();

            return Ok(new
            {
                items
            });
        }
        private static double CalculateDistance(
    double lat1,
    double lon1,
    double lat2,
    double lon2)
        {
            const double R = 6371000;

            var dLat = (lat2 - lat1) * Math.PI / 180;
            var dLon = (lon2 - lon1) * Math.PI / 180;

            var a =
                Math.Sin(dLat / 2) * Math.Sin(dLat / 2) +
                Math.Cos(lat1 * Math.PI / 180) *
                Math.Cos(lat2 * Math.PI / 180) *
                Math.Sin(dLon / 2) * Math.Sin(dLon / 2);

            var c = 2 * Math.Atan2(Math.Sqrt(a), Math.Sqrt(1 - a));

            return R * c;
        }
    }

}